Move domain types to `miden-node-proto`
