+++ block-producer/Cargo.toml
@@ -17,6 +17,7 @@ winterfell = "0.7"
 async-trait = "0.1"
 itertools = "0.11"
 miden-air = { package = "miden-air", git = "https://github.com/0xPolygonMiden/miden-vm.git", branch = "next", default-features = false }
+miden-node-proto = { path = "../proto", version = "0.1"}
 miden_objects = { workspace = true }
 miden_stdlib = { package = "miden-stdlib", git = "https://github.com/0xPolygonMiden/miden-vm.git", branch = "next", default-features = false }
 miden_vm = { package = "miden-vm", git = "https://github.com/0xPolygonMiden/miden-vm.git", branch = "next", default-features = false }
