+++ proto/Cargo.toml
@@ -12,7 +12,6 @@ rust-version = "1.73"
 [dependencies]
 hex = { version = "0.4" }
 miden-crypto = { workspace = true }
-miden-block-producer = { path = "../block-producer", version = "0.1"}
 miden_objects = { workspace = true }
 prost = { version = "0.12" }
 tonic = { version = "0.10" }
